requirements.txt
aiogram>=3.0.0
aiohttp>=3.8.0
python-dotenv>=0.19.0
googletrans==4.0.2
openai>=1.0.0
